requirements.txt
python-binance==1.0.19
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
torch==2.0.1
ta==0.10.2
python-dotenv==1.0.0
websocket-client==1.6.1
